+++ packages/sheets/android/src/main/java/expo/modules/sheets/RNToolsSheetsView.kt
@@ -174,7 +174,7 @@ class RNToolsSheetsView(context: Context, appContext: AppContext) : ExpoView(con
     }
   }
 
-  fun convertToPx(height: Int): Int {
+  fun convertToPx(height: Double): Int {
     val density = context.resources.displayMetrics.density
     return (height * density).toInt()
   }
